Move ui supervise entities service in admin services
