+++ ui/main/src/app/services/admin/SupervisedEntitiesService.ts
@@ -9,8 +9,8 @@
 
 import {map, takeUntil, tap} from 'rxjs/operators';
 import {Observable, Subject} from 'rxjs';
-import {ServerResponseStatus} from '../../server/serverResponse';
-import {ErrorService} from '../error-service';
+import {ServerResponseStatus} from '../../business/server/serverResponse';
+import {ErrorService} from '../../business/services/error-service';
 import {SupervisedEntity} from '@ofServices/admin/model/SupervisedEntity';
 import {SupervisedEntitiesServer} from '@ofServices/admin/server/SupervisedEntitiesServer';
 import {LoggerService as logger} from 'app/services/logs/LoggerService';
